refactor(test): split user degree method tests
